# docs/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: HCML Monitoring Portal (BFF)
  version: 1.0.0
  summary: A read-only backend-for-frontend over the Zabbix 7.0 JSON-RPC API.
  description: |
    The portal owns **no database**. Every response below is either passed through from Zabbix or
    derived in-process and cached in a 500-entry LRU that is lost on restart. Sites, links,
    incidents, host reachability and the monthly SLA have no table anywhere.

    **One write exists** in the whole surface: `POST /api/problems/acknowledge`. It uses a separate
    write-capable Zabbix token and is refused with 503 when that token is unset.

    ### Authentication

    Two schemes, and both work on every guarded path:

    - `Authorization: Bearer <jwt>`, the normal one, from `POST /api/auth/login`. 12-hour expiry.
    - `?token=<jwt>`: a query-string alternative. It exists because `EventSource` cannot set a
      header, but `auth.ts:129-132` reads it on **every** `/api/*` request, not only the SSE ones.
      `logging.ts` redacts it from the request log.

    When `AUTH_ENABLED=false` no route is guarded at all and every caller is treated as `admin`.

    ### Roles

    OpenAPI cannot express a role requirement per path, so each operation states its role in the
    description. From `ROUTE_RULES` (`server/src/auth.ts:35-40`), first match wins, unmatched is
    `viewer`:

    | Role | Paths |
    |---|---|
    | `admin` | `/api/reports/inventory` |
...
    | `viewer` | everything else |

    ### Errors

...

...

...

...
servers:
...
security:
  - bearerAuth: []
  - queryToken: []
paths:
...
  /api/hosts:
    get:
      summary: Every monitored host, with its resolved site
      description: |
        Cached 30 s. `site` is derived by `naming.ts` from four ranked signals: tag, host name,
        inventory, group, because **no HCML host carries a site tag**. It is null when none match.
      operationId: getHosts
      tags: [Hosts]
      responses:
        '200':
          description: Hosts.
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/ZHost'
                    - type: object
                      properties:
                        site:
                          oneOf:
                            - $ref: '#/components/schemas/SiteRef'
                            - type: 'null'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '502': { $ref: '#/components/responses/ZabbixError' }
        '503': { $ref: '#/components/responses/ZabbixUnavailable' }
  /api/hosts/overview:
    get:
      summary: Hosts with reachability state and problem counts
      description: |
        Cached 15 s. `state` is the portal's own reachability judgement (`reachability.ts`), ping
        first, interface flags only as a fallback, on 17 Sep, 16 of 23 hosts Zabbix called
        unavailable answered ping. `reason` records which signal decided it.
      operationId: getHostsOverview
      tags: [Hosts]
      responses:
        '200':
          description: Hosts with derived state.
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/ZHost'
                    - type: object
                      required: [state, reason, problems]
                      properties:
                        state: { $ref: '#/components/schemas/HostState' }
                        reason: { $ref: '#/components/schemas/StateReason' }
                        problems: { $ref: '#/components/schemas/ProblemCounts' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '502': { $ref: '#/components/responses/ZabbixError' }
        '503': { $ref: '#/components/responses/ZabbixUnavailable' }
  /api/hostgroups:
    get:
      summary: Zabbix host groups
      description: Straight `hostgroup.get` passthrough. Cached 60 s. 23 groups in HCML's estate.
      operationId: getHostGroups
      tags: [Hosts]
      responses:
        '200':
          description: Groups.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    groupid: { type: string }
                    name: { type: string }
                  additionalProperties: true
        '401': { $ref: '#/components/responses/Unauthorized' }
        '502': { $ref: '#/components/responses/ZabbixError' }
        '503': { $ref: '#/components/responses/ZabbixUnavailable' }
  /api/items:
    get:
      summary: Items for a host, or by explicit ids
      description: |
        Cached 30 s. `hostid` is required **unless** `itemids` is given.

        `search` is passed to Zabbix with `startSearch: true` deliberately: without it Zabbix runs
        `LIKE '%…%'` and scans all 51,845 items at about 105 ms per call.
      operationId: getItems
      tags: [Hosts]
      parameters:
        - name: hostid
          in: query
          description: Required unless `itemids` is supplied. Numeric.
          schema: { type: string, pattern: '^\d+$' }
        - name: itemids
          in: query
          description: Comma-separated numeric ids, de-duplicated. **At most 50** or 400.
          schema: { type: string }
        - name: search
          in: query
          description: Prefix match on `key_`.
          schema: { type: string }
        - name: graphable
          in: query
          description: Numeric items only. Truthy for exactly `1` or `true`; anything else is false.
          schema: { type: string, enum: ['1', 'true'] }
      responses:
        '200':
          description: Items.
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Item' } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '502': { $ref: '#/components/responses/ZabbixError' }
        '503': { $ref: '#/components/responses/ZabbixUnavailable' }
  /api/latest:
    get:
      summary: Latest values, paged or capped
      description: |
        Cached 15 s. **The presence of `page` switches mode**, and the two modes differ in more than
        pagination: paged mode searches `name` OR `key_`, un-paged searches `name` only.

        With neither `hostid` nor `groupid` this returns an empty result with **200**, not 400.
      operationId: getLatest
      tags: [Hosts]
      parameters:
        - name: hostid
          in: query
          schema: { type: string, pattern: '^\d+$' }
        - name: groupid
          in: query
          schema: { type: string, pattern: '^\d+$' }
        - name: search
          in: query
          schema: { type: string }
        - name: page
          in: query
          description: Presence switches to paged mode.
          schema: { type: integer, minimum: 1, maximum: 1000000 }
        - name: pageSize
          in: query
          schema: { type: integer, minimum: 25, maximum: 500, default: 100 }
      responses:
        '200':
          description: |
            `total`, `page` and `pageSize` are present only in paged mode. `truncated` means the cap
            was hit: 500 rows un-paged, 20,000 ids paged.
          content:
            application/json:
              schema:
                type: object
                required: [items, truncated]
                properties:
                  items: { type: array, items: { type: object, additionalProperties: true } }
                  truncated: { type: boolean }
                  total: { type: integer }
                  page: { type: integer }
                  pageSize: { type: integer }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '502': { $ref: '#/components/responses/ZabbixError' }
        '503': { $ref: '#/components/responses/ZabbixUnavailable' }
  /api/problems:
    get:
      summary: Open problems
      description: |
        Cached 5 s, and **never served stale**: an operator who has just acknowledged must not be
        shown the previous list. About 664 problems at HCML, 459 kB compressing to 25 kB.
      operationId: getProblems
      tags: [Problems]
      responses:
        '200':
          description: Open problems, newest first.
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Problem' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '502': { $ref: '#/components/responses/ZabbixError' }
        '503': { $ref: '#/components/responses/ZabbixUnavailable' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: The normal scheme. 12-hour expiry, carries `{ sub, role }`.
    queryToken:
      type: apiKey
      in: query
      name: token
      description: |
        The same JWT in the query string. Required for `EventSource`, which cannot set headers, but
        accepted on **every** guarded path. Kept out of the request log by `logging.ts`.
  responses:
    BadRequest:
      description: '`bad_request`, a parameter failed validation.'
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unauthorized:
      description: '`unauthorized`, missing or expired portal JWT. Not the same as `zabbix_auth`.'
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    ZabbixError:
      description: '`zabbix_error`, unreachable, non-2xx, non-JSON-RPC, or a JSON-RPC error.'
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    ZabbixUnavailable:
      description: |
        `zabbix_timeout` (no answer within `ZABBIX_TIMEOUT_MS`, default 10 s) or `zabbix_auth` (Zabbix
        rejected the API token). Both are configuration states rather than upstream faults, which is
        why they are 503 and not 504.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  schemas:
    Error:
      type: object
      required: [error, message]
      description: The only error shape this server produces, at every status.
      properties:
        error: { type: string, description: A stable machine-readable code. }
        message: { type: string, description: Safe to show a user. Never contains the Zabbix URL. }
...
    SiteRef:
      type: object
      description: A resolved site. No Zabbix entity corresponds to this.
      properties:
        code: { type: string }
        name: { type: string }
      additionalProperties: true
    HostState:
      type: string
      enum: [up, down, degraded, nodata, disabled]
      description: The portal's reachability judgement, ping first.
    StateReason:
      type: string
      enum: [ping, interface, snmp-silent, agent-silent, stale, no-interface, unsupported, disabled]
      description: Which signal decided `HostState`. Provenance, so a wrong call can be traced.
    ProblemCounts:
      type: object
      required: [total, bySeverity]
      properties:
        total: { type: integer }
        bySeverity: { type: object, additionalProperties: { type: integer } }
    ZHost:
      type: object
      properties:
        hostid: { type: string }
        name: { type: string }
        status: { type: string }
        maintenance_status: { type: string }
        interfaces:
          type: array
          items: { type: object, additionalProperties: true }
      additionalProperties: true
    Item:
      type: object
      properties:
        itemid: { type: string }
        hostid: { type: string }
        name: { type: string }
        key_: { type: string }
        value_type: { type: string }
        units: { type: string }
        delay: { type: string }
        status: { type: string }
        state: { type: string }
        lastvalue: { type: string }
        lastclock: { type: string }
        tags: { type: array, items: { $ref: '#/components/schemas/Tag' } }
      additionalProperties: true
    Tag:
      type: object
      required: [tag, value]
      properties:
        tag: { type: string }
        value: { type: string }
    Problem:
      type: object
      required: [eventid, objectid, object, name, severity, clock, acknowledged]
      properties:
        eventid: { type: string }
        objectid: { type: string }
        object: { type: string, description: '`0` is a trigger.' }
        name: { type: string }
        severity: { type: string, enum: ['0', '1', '2', '3', '4', '5'] }
        clock: { type: string }
        r_eventid: { type: string, description: '`0` while still open.' }
        r_clock: { type: string }
        acknowledged: { type: string, enum: ['0', '1'] }
        suppressed: { type: string, enum: ['0', '1'] }
        opdata: { type: string }
        tags: { type: array, items: { $ref: '#/components/schemas/Tag' } }
        host: { type: string, description: Enriched by the BFF. }
        hostid: { type: string, description: Enriched by the BFF. }
        manualClose:
          type: boolean
          description: Whether the trigger permits manual close. Gates the Close button.
      additionalProperties: true
...